Restructured client, server and logging packages. Configured logging. Added a main example.go file
